+++ modules/ai/index.mjs
@@ -1,0 +1,19 @@
+/*
+ * Copyright (C) 2015, 2025; Green Screens Ltd.
+ */
+
+export * from "./GSAI.mjs";
+
+/*
+NOTE: This module is a work in progress and is not yet fully functional.
+
+* There are few more Browser AI API in preview mode which will be implemented later.
+
+* Higher level functions might be available in the future components
+  like GSAITranslator, GSAIDetector, GSAISummarizer.
+  
+* The plan is to add AI features to form inputs, and textareas for advanced text processing.  
+
+* Check WebLLM project for more, and examples: https://codepen.io/neetnestor/pen/vYwgZaG
+
+*/