+++ modules/base/index.mjs
@@ -35,3 +35,4 @@ export * from "./GSRouter.mjs";
 export * from "./GSTemplateCache.mjs";
 export * from "./GSTheme.mjs";
 export * from "./GSUtil.mjs";
+